jclklib: Move all API related definition to jclk_init.hpp
Move all API related definition to jclk_init.hpp so that user
application will just need to include one header file.
In future, this file will be move to pub folder.
