+++ jclklib/client/Makefile
@@ -15,7 +15,7 @@ JCLKLIB_TOPLEVEL_DIR = $(JCLKLIB_CLIENT_DIR)/..
 JCLKLIB_COMMON_DIR = $(JCLKLIB_TOPLEVEL_DIR)/common
 LIBS = pthread rt
 LIBS_FLAGS = $(foreach lib,$(LIBS),-l$(lib))
-OBJ = init msgq_tport message null_tport transport msgq_tport notification_msg connect_msg client_state subscribe_msg
+OBJ = jclk_init msgq_tport message null_tport transport msgq_tport notification_msg connect_msg client_state subscribe_msg
 COMMON_OBJ = print sighandler transport msgq_tport message connect_msg subscribe_msg notification_msg jclklib_import
 OBJ_FILES = $(foreach f,$(OBJ),$(f).o)
 DEPENDS = $(foreach f,$(OBJ),$(f).d)
